Add zipping tool
Add a tool to take a set of migrated downstream refs and zip them,
interleaving commits as directed by a provided "umbrella" branch.  The
umbrella is expected to be a series of submodule update commits,
possibly interspersed with commits that do not update submodules.
Commits that do not update submodules are ignored.

For each commit updating submodules, the tool rewrites the commit with
the changes in the corresponding migrated downstream commits and takes
care to preserve merges from upstream commits.  The end result is a
branch that has all of the sumodule updates "inlined" into a single
history.