Ignore bad submodule updates by default
Sometimes a submodule update refers to a commit that no longer exists.
Typically, this is because the user made an error and included a
submodule update in a commit unintentionally.

With this change, such updates are ignored by default.  Add an option
--abort-bad-submodule to cause the script to terminate in such a
situation.  This replaces the code that used to query the user for
what to do.  Doing things this way makes zip-downstream-fork.py more
easily scriptable.