Preserve upstream trees
If an upstream project doesn't exist under the umbrella repository,
import its contents from the upstream monorepo.

This takes some care to ensure the upstream bits are in sync with the
submodule updates.  The upstream parts should never be updated past
the point of the most recent merge from upstream via a submodule
update.  Each submodule update is checked for merges from upstream and
the latest merged commit is used as the base of the zipped commit
tree.  If there are no upstream merges, the base is taken from the
base tree of the parent zipped commit.  If there are multiple parents,
the base tree of each is examined and the latest one according to
commit topological order is taken as the base tree.