Inline first umbrella commit if upstream submodule add
If the very first commit in the umbrella adds a single submodule
at a commit from upstream, just return the upstream commit for the
umbrella commit.  This prevents unwanted commit duplication.  If
we rewrote the umbrella commit it would have the same content as
the upstream commit build would have a different parent set.