Don't unnecessarily reattribute commits
If a commit only updates one submodule, use the author, committer
and date information from the submodule's commit since that's the
content that's being inlined.

If a commit updates multiple submodules, then it seems appropriate
to use the author, committer and date information from the umbrella
commit, since it was a deliberate coupling of subproject commits.