Do not rename tags
Whether or not to rename individual tags is a hard problem.  Sometimes
a commit wants a newly-named tag and sometimes it wants the existing
tag updated.  Let the user do any tag renaming manually.  Add an
--update-tags option to let the user say whether to not to rewrite
tags.  All tags will be migrated to new commits with that option.  If
that's not wanted, the user will have to migrate individual tags
manually.