Only check parent history when valid
If a parent commit is from the umbrella, it may only have a mark and
git-fast-filter may not have written out the commit yet.  Thus we
can't check ancestry on it.  Keep a map from new umbrella commits to
old submodule commits and check ancestry against the old commits if an
umbrella parent was originally a submodule commit (i.e. it was inlined
into the umbrella).

Remember which commits belong to the umbrella repository and skip them
for ancestry checking purpose.  We can also skip checking ancestry
against upstream monorepo commits because the tracked merge base
should already include submodule commits.  If it doesn't it means we
erred in calculating the merge base.