Track previous submodule state along branches
Keep a record of the previous submodule state for each umbrella
commit.  At merge points, see if the submodule state along any parent
is equal to the submodule state of the current commit.  If it is, the
corresponding submodule was not updated.  Otherwise, it was.

With this change we handle branches within the umbrella history and
know whether merge commits updated any submodules.