Track inlined submodules more accurate for tag rewriting
Track exactly which submodule trees changed in each umbrella commit so
we know where to map tags from submodule commits.  If a submodule
commit was inlined into an umbrella commit, its tag should point to
the new zipped commit.  If a submodule commit is simply a parent of a
zippped commit, its tag should be updated to the new submodule commit
(rewritten because its parents were rewritten).