Fix using migrate-downstream-fork with up-to-date monorepo
Commit 2f2aeb9b has the following in its log:

llvm-svn: 181277
llvm-svn: 181274
llvm-svn: 181273

I assume this is because it is a commit that exists in 3 subprojects, and
thus has 3 separate svn revisions which were combined during monorepo
conversion.

find_svnrev assumes a 1-to-1 git commit to svn revision correspondance
which seems inaccurate with how the monorepo is now laid out.

This commit now has the function return a list of tuples, so that we can
represent the 1-to-many git commit to svn revision relationship.

This also required a change to the regex, which used '\Z' to match the
assumed 1-to-1 svn revision associated with the commit. This assumption is
no longer true in the monorepo. The new strategy is to start at the end of
the message and work our way up, matching each line as an svn revision
until we reach the first non-match.