migrate-downstream-fork: fixing souce-kind usage of 'split'
Only in actual argument description "split" source-kind is spelled as 'split'.
Every other place spells it as 'merge-split'. It does not affect autodetect
execution but leads to unpleasant effect when trying to specify 'split' manually.

Since in https://reviews.llvm.org/rL353713 source-kind was mentioned as 'split',
I figured that fixing it to 'split' spelling is a right way to go.